Modify the build process to have separate compilation of the esternal program 'interface'.
