+++ src/CMakeLists.txt
@@ -41,4 +41,4 @@ set_target_properties(iqs
     POSITION_INDEPENDENT_CODE 1
 )
 
-
+#add_subdirectory(interface)
